backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
cryptography==41.0.7
celery==5.3.6
redis==5.0.1
httpx==0.25.2
lxml==4.9.3
python-dateutil==2.8.2
openpyxl==3.1.2
aiofiles==23.2.1
slowapi==0.1.9
python-dotenv==1.0.0
requests==2.31.0
